[MOB] ADD: support multilingual #43
