+++ assets/translations/ko.json
@@ -9,7 +9,7 @@
   "year": "년",
   "month": "월",
   "day": "일",
-
+  "permission_request": "권한을 허용해주세요.",
 
   "header.upcoming": "다가오는 편지",
   "header.d_day": "D-{day}",
