+++ ios/ko.lproj/InfoPlist.strings
@@ -1,0 +1,4 @@
+"NSCameraUsageDescription" = "사진 촬영을 위해 카메라 접근이 필요합니다.";
+"NSMicrophoneUsageDescription" = "음성 녹음을 위해 마이크 접근이 필요합니다.";
+"NSPhotoLibraryUsageDescription" = "편지에 사진을 첨부하기 위해 사진 앨범에 접근합니다. 선택한 사진은 편지와 함께 저장되며, 본인만 열람할 수 있습니다.";
+"NSUserTrackingUsageDescription" = "맞춤형 광고를 제공하기 위해 사용자의 앱 내 활동 정보를 수집하여 광고 콘텐츠를 사용자에게 적합하게 조정합니다.";
