+++ ios/ko.lproj/LaunchScreen.strings
@@ -1,0 +1,1 @@
+
