+++ ios/ko.lproj/Main.strings
@@ -1,0 +1,1 @@
+
