+++ lib/widgets/common/image_upload.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:naro/controllers/image_upload_controller.dart';
 import 'package:naro/styles/colors.dart';
-import 'package:naro/utils/permisson_manager.dart';
+import 'package:naro/utils/permission_manager.dart';
 import 'package:naro/widgets/common/image_viewer.dart';
 import 'package:flutter/services.dart';
 import 'package:easy_localization/easy_localization.dart';
@@ -69,10 +69,13 @@ class _PhotoUploadState extends State<ImageUpload> {
       onTap: () async {
         HapticFeedback.lightImpact();
         final isGranted = await PermissionManager().requestCameraPermission(context);
+        print('Image upload: $isGranted');
         if (!isGranted) {
           return;
         }
+        print('Image upload button tapped1');
         final picked = await _picker.pickMultiImage(limit: 3);
+        print('Image upload button tapped2');
         if (picked.isNotEmpty) {
           setState(() {
             widget.imageController.addImages(picked);
